+++ src/commands/pipeline.ts
@@ -1,17 +1,18 @@
 import { Project } from '../common/Config';
 import { PipelineCommand } from '../common/pipelines';
 import { createPipeline, Response, StatusCode } from '../common/api';
+import { IPipeline } from '../common/iPipeline';
 
 export class Pipeline extends PipelineCommand {
-  protected async runPerProject<T>(project: Project): Promise<Response<T>> {
+  protected async runPerProject(project: Project) {
     const resp = await this.triggerPipeline(project);
     if (resp.status === StatusCode.Success && this.yargs.await) {
-      return await this.awaitPipelineCompletion(project, this.yargs.ref);
+      return await this.awaitPipelineCompletion(project, resp.data);
     }
     return resp;
   }
 
-  private triggerPipeline(project: Project): Promise<Response<any>> {
+  private triggerPipeline(project: Project): Promise<Response<IPipeline>> {
     return createPipeline(this.config.uri, project.id, this.yargs.ref).then(
       data => {
         this.screenPrinter.setProjectSuccess(project, 'Pipeline crated');